refactor(lld/manager): rename render prop
